Converted Evaluator to returning Tokens instead of strings
Made internal Lexeme array in Evaluator into an list.
Removed index and offset variables in Evaluator in favor of removing the lexemes from the lexeme list once evaluated into a token.
Set groundwork for having classes for each token with each token class inheriting from Token.
Simplified GetNextLexeme by fixing an edge case when Chars.Count was <= 1.
Added NewLine to the CharType enum to signify the end of a statment.
Improved initial program parsing to replace newlines with unicode new lines so that newlines can be preserved in when put into the char array.